dag: Show a loading message when clicking commits
The diff_info() task can take a long time so use a background task and
show a temporary loading message.

For whatever reason (perhaps the GIL?) the GUI still blocks while the
task is running.  That's unfortunate; QRunnable was supposed to avoid
that.  Maybe we can try using QThread instead.

Signed-off-by: David Aguilar <davvid@gmail.com>