main.view: Use a better default layout
Rework the original widget sizes and layouts to more closely match how
the tool is intended to be used.  For example, we now hide the Console
and Action tools by default since they are not the main focus of
interaction.

We use a trick to resize dockwidgets -- we temporarily set a maximum
size on the child widgets so that Qt uses this as their initial size.
This is not done when restoring a layout.

Increment the widget_version to invalidate old layouts.  This ensures
that everyone will get the new layout.

Signed-off-by: David Aguilar <davvid@gmail.com>