main.view: Avoid common keyboard input sequences
Some keyboard layouts, for example the standard British layout, use
Alt+$number and Shift+Alt+$number to input common characters such as €
and #.

Avoid the problem by switching the default shortcuts to use the Ctrl key
instead of Alt.

OS X caveat: Ctrl means "command" there, and that overlaps with the
"switch spaces" keybindings, so we use Meta aka "control" there.
This is counter to all the other keybindings which use the "command"
key whenever the documentation mentions "Ctrl".

Longer-term, it would be nice if we had a GUI for configuring keyboard
shortcuts so that users could configure the tool to their liking.

Closes #146

Signed-off-by: David Aguilar <davvid@gmail.com>